proxy: pre-release versions support
When current module version is a pre-release, then fetch all versions
including pre-releases from the Go proxy and try to upgrade. This
includes both stable and unstable versions.

But if the current module version is a stable version, then exclude
pre-release versions from the list of available versions.

This behavior also means that if a current version is pinned to a
pre-release, gobump will keep upgrading pre-releases until the first
stable version is reached. After that, only stable versions will be
considered.